backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.36
pydantic==2.9.0
pydantic-settings==2.5.2
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-text-splitters==0.3.0
faiss-cpu==1.8.0
huggingface-hub==0.25.2
pypdf==5.0.1
requests==2.32.3
